% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonoscopy.R
\docType{package}
\name{colonoscopy-package}
\alias{colonoscopy}
\alias{colonoscopy-package}
\title{Exposing or hiding package dependencies in R code.}
\description{
colonoscopy allows you to transform R code by adding or
removing namespace access operators (`::` and `:::`) in accordance
with the search path in an environment.
}
\details{
RStudio users can find addins "Scope selection" and "Unscope selection"
to apply the respective function and replace the selected code.
}
\examples{
scope("1+1")
scope("tail")
scope("scope")
scope(scope)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TylerGrantSmith/colonoscopy}
  \item Report bugs at \url{https://github.com/TylerGrantSmith/colonoscopy/issues}
}

}
\author{
\strong{Maintainer}: Tyler Grant Smith \email{tyler.grant.smith@gmail.com}

}
