% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Remove all namespace accessors from an expression}
\usage{
pack(x, envir = NULL)
}
\arguments{
\item{x}{An expression to process.  Input is automatically quoted, use !! to unquote if
you have already captured an expression object.}

\item{envir}{An environment in which the expression should be evaluated.}
}
\description{
Remove all namespace accessors from an expression
}
\examples{
# Unchanged
pack(1)
pack("a")

# Modified
pack(packr::pack)
}
