% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack_function}
\alias{pack_function}
\title{Remove all namespace accessors from a function definition}
\usage{
pack_function(f, enclos = environment(f))
}
\arguments{
\item{f}{A function}

\item{enclos}{The environment in which the function is defined.  Defaults to
\code{environment(f)}}
}
\description{
Remove all namespace accessors from a function definition
}
\examples{
test_function <- function() { packr::pack }
pack_function(test_function)
}
