% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packr.R
\docType{package}
\name{packr-package}
\alias{packr}
\alias{packr-package}
\title{packr: packr}
\description{
Utility to rewrite function calls to make them use double colon and triple colon operators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TylerGrantSmith/packr}
  \item Report bugs at \url{https://github.com/TylerGrantSmith/packr/issues}
}

}
\author{
\strong{Maintainer}: Tyler Grant Smith \email{tyler.grant.smith@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
