% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{scope}
\alias{scope}
\alias{scope.default}
\alias{scope.character}
\alias{scope.function}
\title{Make package dependencies explicit}
\usage{
scope(x, envir = caller_env(), ...)

\method{scope}{default}(x, envir = caller_env(), ...)

\method{scope}{character}(x, envir = caller_env(), ...)

\method{scope}{`function`}(x, envir = get_env(x) \%||\% caller_env(), useSource = TRUE, ...)
}
\arguments{
\item{x}{A character vector of parseable code or a function.}

\item{envir}{An environment in which the expression should be evaluated.
Defaults to the enclosing environment if \code{x} is a function, otherwise the calling environment.}
}
\value{
Returns a [ParseData] object which prints the modified code.
}
\description{
Parses the input, and explicitly adds namespace access where appropriate.
}
\examples{

# Unchanged
scope(1)
scope("a")

# Modified
library(colonoscopy)
scope(scope)
}
\keyword{internal}
