% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack_function}
\alias{unpack_function}
\title{Make explicit package dependencies in a function definition}
\usage{
unpack_function(f, enclos = environment(f))
}
\arguments{
\item{f}{A function}

\item{enclos}{The environment in which the function is defined.  Defaults to
\code{environment(f)}}
}
\description{
Make explicit package dependencies in a function definition
}
\examples{
unpack_function(unpack_function)
}
