% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscope.R
\name{unscope}
\alias{unscope}
\title{Remove \code{`::`} and \code{`:::`} operators}
\usage{
unscope(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Deletes all package access operators.
}
\examples{

# No changes.
unscope("2+2")

# Removes "utils::" and "colonoscopy::".
unscope("utils::head(colonoscopy::scope)")
}
