% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonoscopy.R
\docType{package}
\name{colonoscopy-package}
\alias{colonoscopy}
\alias{colonoscopy-package}
\title{colonoscopy: Tools for adding and removing namespace access colonoperators, `::` and `:::`.}
\description{
Add something later.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TylerGrantSmith/colonoscopy}
  \item Report bugs at \url{https://github.com/TylerGrantSmith/colonoscopy/issues}
}

}
\author{
\strong{Maintainer}: Tyler Grant Smith \email{tyler.grant.smith@gmail.com}

}
\keyword{internal}
