% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\alias{unpack.default}
\alias{unpack.character}
\alias{unpack.function}
\title{Make explicit package dependencies}
\usage{
unpack(x, envir, ...)

\method{unpack}{default}(x, envir, ...)

\method{unpack}{character}(x, envir = caller_env(), ...)

\method{unpack}{`function`}(x, envir = caller_env(), use_fn_env = TRUE, useSource = T, ...)
}
\arguments{
\item{x}{An expression to process.  Input is automatically quoted, use !! to unquote if
you have already captured an expression object.}

\item{envir}{An environment in which the expression should be evaluated.}
}
\description{
\code{unpack} returns an expression with all symbols resolving to loaded namespaces by
explicitly adding the namespace access notation.
}
\examples{

# Unchanged
unpack(1)
unpack("a")

# Modified
library(packr)
unpack(unpack)
}
