% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscope.R
\name{unscope}
\alias{unscope}
\title{Remove \code{`::`} and \code{`:::`} operators}
\usage{
unscope(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
TBD
}
\examples{

unscope("2+2")

unscope("utils::head(colonoscopy::scope)")
}
