% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscope.R
\name{unscope}
\alias{unscope}
\title{Remove \code{`::`} and \code{`:::`} operators}
\usage{
unscope(x, ...)
}
\arguments{
\item{x}{A character vector of parseable R code, or a function.}

\item{...}{Optional arguments when x is a function.}
}
\description{
Deletes all package access operators.
}
